CAN sending cleanup to separate files
